Generate random password for internal user
If the user in elgg is created with the same password as the one in LDAP and later the LDAP password changes, the user is able to log on with the new LDAP password or the old one via internal authentication. Creating a random complicated password sorts that out as the internal password is not known to anybody.